% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addBackticks}
\alias{addBackticks}
\title{Add backticks to a string.}
\usage{
addBackticks(str)
}
\arguments{
\item{str}{A string.}
}
\value{
String surrounded by backticks.
}
\description{
Add backticks at the beginning and end of a string. This avoids situation
where lazy evaluation fails due to column names which are not legal as R
variable names.
}
