% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_cytometry.R
\name{auroraPreprocess}
\alias{auroraPreprocess}
\title{Preprocess Aurora sample.}
\usage{
auroraPreprocess(sample, cofactor = 150)
}
\arguments{
\item{sample}{An Astrolabe sample.}

\item{cofactor}{Cofactor for asinh transformation.}
}
\value{
Sample after the above steps are done.
}
\description{
Quality control, cleaning, and transformation for Aurora sample.
}
