% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_cytometry.R
\name{auroraTransformChannels}
\alias{auroraTransformChannels}
\title{Transform Aurora channels.}
\usage{
auroraTransformChannels(flow_frame)
}
\arguments{
\item{flow_frame}{FlowCore flow frame.}
}
\value{
Transformed flow_frame.
}
\description{
Given a FlowCore flow frame, transform channel values using estimateLogicle.
}
