% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs.R
\name{calculateFcsDigest}
\alias{calculateFcsDigest}
\title{Calculate parameter digest for sample.}
\usage{
calculateFcsDigest(sample, parameter_name = NULL)
}
\arguments{
\item{sample}{An Astrolabe sample.}

\item{parameter_desc, parameter_name}{Parameter descriptions and names from
an FCS file.}
}
\value{
Parameter digest for data.
}
\description{
Apply the digest::digest function to the list of parameter descriptions and
parameter names for a given FCS data structure.
}
