% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citeseq.R
\name{citeseqPreprocess}
\alias{citeseqPreprocess}
\title{Preprocess citeseq sample.}
\usage{
citeseqPreprocess(sample, cofactor = 5)
}
\arguments{
\item{sample}{An Astrolabe sample.}

\item{cofactor}{Cofactor for asinh transformation.}
}
\value{
Sample after the above steps are done.
}
\description{
Quality control, cleaning, and transformation for citeseq sample.
}
