% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs.R
\name{convertFlowFrame}
\alias{convertFlowFrame}
\title{Convert a FlowCore flow_frame to Astrolabe sample.}
\usage{
convertFlowFrame(experiment, filename, flow_frame)
}
\arguments{
\item{experiment}{An Astrolabe experiment.}

\item{filename}{The name of the FCS file to import.}

\item{flow_frame}{FlowCore flow frame.}
}
\value{
FCS data, in orloj internal FCS list format.
}
\description{
Convert the flow_frame class into orloj's internal FCS list format.
}
\details{
The orloj FCS list format will accumulate more fields as analyses are
applied to it. For example, pre-processing will add a mask to find the
non-bead indices. You can use \code{\link{fcsExprs}} the get the expression
data after applying all of the masks that are in the list.
}
\seealso{
\code{\link{isSample}}, \code{\link{fcsExprs}}
}
