% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{differentialAbundanceAnalysis}
\alias{differentialAbundanceAnalysis}
\title{Differential abundance analysis.}
\usage{
differentialAbundanceAnalysis(
  experiment,
  level = .chooseLevel(experiment),
  convert_ids = TRUE
)
}
\arguments{
\item{experiment}{An Astrolabe experiment.}

\item{level}{Cell subset level, such as "Assignment" or "Profiling".}

\item{convert_ids}{Whether to convert Astrolabe IDs to feature names.}
}
\value{
Differential abundance analysis list.
}
\description{
Load the experiment differential abundance analysis, for a given cell subset
level.
}
