% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experimentAssignmentSubsets}
\alias{experimentAssignmentSubsets}
\title{List of Assignment subsets.}
\usage{
experimentAssignmentSubsets(experiment)
}
\arguments{
\item{experiment}{An Astrolabe experiment.}
}
\value{
Character vector of Assignment subsets.
}
\description{
Return a vector of all Assignment subsets.
'
}
