% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experimentCellSubsetChannelStatistics}
\alias{experimentCellSubsetChannelStatistics}
\title{Experiment cell subset channel statistics.}
\usage{
experimentCellSubsetChannelStatistics(experiment,
  level = .chooseLevel(experiment))
}
\arguments{
\item{experiment}{An Astrolabe experiment.}

\item{level}{Cell subset level. Currently supported levels are "Assignment"
and "Profiling". The default is "Profiling" for Profiling Only experiments
and "Assignment" otherwise.}
}
\value{
Cell subset channel statistics for that level.
}
\description{
Cell subset channel statistics for all of the samples in an experiment.
}
