% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experimentCellSubsetCounts}
\alias{experimentCellSubsetCounts}
\title{Experiment cell subset counts.}
\usage{
experimentCellSubsetCounts(experiment, level = "Assignment")
}
\arguments{
\item{experiment}{An Astrolabe experiment.}

\item{level}{Cell subset level. Currently supported levels are "Assignment"
and "Profiling".}
}
\value{
Cell subset counts for that level.
}
\description{
Cell subset counts for all of the samples in an experiment.
}
