% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experimentDifferentialExpressionAnalysis}
\alias{experimentDifferentialExpressionAnalysis}
\title{Differential expression analysis.}
\usage{
experimentDifferentialExpressionAnalysis(experiment, convert_ids = TRUE)
}
\arguments{
\item{experiment}{An Astrolabe experiment.}

\item{convert_ids}{Whether to convert Astrolabe IDs to user-supplied values.}
}
\value{
Differential expression analysis list.
}
\description{
Load the experiment differential expression analysis across all levels and
all kits.
}
