% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experimentDifferentialExpressionAnalysisValueVar}
\alias{experimentDifferentialExpressionAnalysisValueVar}
\title{Differential expression analysis value var.}
\usage{
experimentDifferentialExpressionAnalysisValueVar(experiment)
}
\arguments{
\item{experiment}{An Astrolabe experiment.}
}
\value{
Value variable.
}
\description{
Return the name of the value variable that was used for differential
expression analysis in this experiment.
}
