% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experimentLabelingHierarchy}
\alias{experimentLabelingHierarchy}
\title{Experiment labeling hierarchy.}
\usage{
experimentLabelingHierarchy(experiment)
}
\arguments{
\item{experiment}{An Astrolabe experiment.}
}
\value{
User-readable table of the experiment hierarchy.
}
\description{
Return a user-readable table of the experiment labeling hierarchy, listing
the strategy for all terminal Assignment nodes.
}
