% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experimentLabelingStrategy}
\alias{experimentLabelingStrategy}
\title{Experiment Assignment labeling strategy.}
\usage{
experimentLabelingStrategy(experiment)
}
\arguments{
\item{experiment}{An Astrolabe experiment.}
}
\value{
User-readable table of the experiment hierarchy.
}
\description{
Return a user-readable table of the labeling strategy for the Assignment
level.
}
