% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experimentMds}
\alias{experimentMds}
\title{MDS map.}
\usage{
experimentMds(experiment, level = .chooseLevel(experiment),
  convert_ids = TRUE)
}
\arguments{
\item{experiment}{An Astrolabe experiment.}

\item{level}{Cell subset level, such as "Assignment" or "Profiling".}

\item{convert_ids}{Whether to convert Astrolabe IDs to feature names.}
}
\value{
MDS map.
}
\description{
Return the MDS map for the experiment, for a given level.
}
