% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experimentQcMetrics}
\alias{experimentQcMetrics}
\title{Quality control metrics.}
\usage{
experimentQcMetrics(experiment)
}
\arguments{
\item{experiment}{An Astrolabe experiment.}
}
\value{
List of different quality control metrics.
}
\description{
Load the experiment quality control metrics.
}
