% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experimentSummary}
\alias{experimentSummary}
\title{Astrolabe experiment summary.}
\usage{
experimentSummary(experiment)
}
\arguments{
\item{experiment}{An Astrolabe experiment.}
}
\description{
Astrolabe experiment summary.
}
