% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportImmport.R
\name{exportImmport}
\alias{exportImmport}
\title{Immport export.}
\usage{
exportImmport(file_path, experiment, sample_id)
}
\arguments{
\item{file_path}{File path for the exported tsv file.}

\item{experiment}{An Astrolabe experiment.}

\item{sample_id}{An Astrolabe sample ID.}
}
\description{
Export an Immport CyTOF_derived_data.tsv file for a single sample.
}
