% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{exportReport}
\alias{exportReport}
\title{Export an orloj report.}
\usage{
exportReport(dir, report, file_format = "png", dpi = 100,
  verbose = FALSE)
}
\arguments{
\item{dir}{The name of the dir where figures are to be exported.}

\item{report}{An orloj report.}

\item{file_format}{Exported plots file format (such as png, svg, etc.).}

\item{dpi}{Plot resolution. Only applies only to raster output types.}

\item{verbose}{Whether to provide updates to console.}
}
\description{
Exports an orloj report to a destination directory. Reports take orloj data
structures and generate a series of plots that summarize, illustrate, or
explain an analysis.
}
