% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{findStandardMassCytometryChannels}
\alias{findStandardMassCytometryChannels}
\title{Find indices of standard mass cytometry channels.}
\usage{
findStandardMassCytometryChannels(sample)
}
\arguments{
\item{sample}{An Astrolabe sample.}
}
\value{
A list with indices for each of the standard channels.
}
\description{
Find the indices of a set of standard mass cytometry channels, including
Time, Event Length, Ir191/193 (DNA), and Live/Dead.
}
