% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_cytometry.R
\name{flowPreprocess}
\alias{flowPreprocess}
\title{Preprocess flow cytometry sample.}
\usage{
flowPreprocess(sample)
}
\arguments{
\item{sample}{An Astrolabe sample.}
}
\value{
Sample after the above steps are done.
}
\description{
Quality control, cleaning, and transformation for flow cytometry sample.
}
