% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{formatPvalue}
\alias{formatPvalue}
\title{Format a p-value or FDR into a nice string.}
\usage{
formatPvalue(v, label = "p", digits = 4)
}
\arguments{
\item{v}{A p-value or FDR value to format.}

\item{label}{String describing the value being presented.}

\item{digits}{Number of decimal places to be used.}
}
\value{
A string of a nicely formatted value.
}
\description{
Unlike the R \code{\link{format.pval}}, this function refrains from using
scientific notation.
}
