% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCombineAggregateStatistics.R
\name{getCellCounts}
\alias{getCellCounts}
\title{Get cell counts for a given analysis type.}
\usage{
getCellCounts(aggregate_statistics, analysis)
}
\arguments{
\item{aggregate_statistics}{List of data frames with aggregated statistics,
combined over all samples.}

\item{analysis}{Type of analysis for which to get count data.}
}
\value{
A data fram with cell counts.
}
\description{
Given an analysis type (such as "Assignment" or "Profiling") get the cell
counts for the respective cell subsets.
}
