% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{getCellSubsetLevels}
\alias{getCellSubsetLevels}
\title{Astrolabe cell subset levels.}
\usage{
getCellSubsetLevels(sample)
}
\arguments{
\item{sample}{An Astrolabe sample.}
}
\value{
A data frame with cell subset levels.
}
\description{
Return a data frame where each row corresponds to one level in the cell
subset classification. This includes levels that come from the hierarchy,
an "Assignment" level, and a terminal "Profile" level.
}
