% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{getExperimentSampleFeatures}
\alias{getExperimentSampleFeatures}
\title{Get experiment sample features.}
\usage{
getExperimentSampleFeatures(experiment)
}
\arguments{
\item{experiment}{An Astrolabe experiment.}
}
\value{
Experiment sample features.
}
\description{
Get the sample features for a given Astrolabe experiment. Feature names are
converted from internal Astrolabe IDs to user-readable names.
}
