% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{getSampleFilename}
\alias{getSampleFilename}
\title{Get sample filename.}
\usage{
getSampleFilename(experiment, sample_id)
}
\arguments{
\item{experiment}{An Astrolabe experiment.}

\item{sample_id}{Astrolabe sample ID.}
}
\value{
The sample's filename.
}
\description{
Get sample filename.
}
