% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{getSampleName}
\alias{getSampleName}
\title{Get sample name.}
\usage{
getSampleName(experiment, sample_id)
}
\arguments{
\item{experiment}{An Astrolabe experiment.}

\item{sample_id}{Astrolabe sample ID.}
}
\value{
The sample's name, as supplied by user.
}
\description{
Get sample name.
}
