% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs.R
\name{identifyFcsInstrument}
\alias{identifyFcsInstrument}
\title{Identify FCS file instrument.}
\usage{
identifyFcsInstrument(flow_frame)
}
\arguments{
\item{flow_frame}{FlowCore flow frame.}
}
\value{
Instrument for flow frame.
}
\description{
Read the CYT field from the FCS file header.
}
