% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citeseq.R
\name{importCiteseqChannels}
\alias{importCiteseqChannels}
\title{Import CITEseq Channel Information.}
\usage{
importCiteseqChannels(filename)
}
\arguments{
\item{filename}{The name of the CITEseq file to import.}
}
\value{
A dataframe with channel name and description.
}
\description{
Import the antibody portion of CITEseq file.
}
