% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs.R
\name{importFcsChannels}
\alias{importFcsChannels}
\title{Import FCS Channel Information.}
\usage{
importFcsChannels(filename)
}
\arguments{
\item{filename}{The name of the FCS file to import.}
}
\value{
A dataframe with channel names and descriptions.
}
\description{
Import the TEXT section of an FCS file and extract channel names and
descriptions (the $PnN and $PnS fields, respectively).
}
