% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs.R
\name{importFcsFile}
\alias{importFcsFile}
\title{Import an FCS File.}
\usage{
importFcsFile(filename, transformation = "linearize")
}
\arguments{
\item{filename}{The name of the FCS file to import.}

\item{transformation}{Which flowCore transformation to use. See
\code{\link[flowCore]{read.FCS}} for more information}
}
\value{
FCS data, in orloj internal FCS list format.
}
\description{
Imports an FCS file using flowCore::read.FCS and convert the flow_frame class
into orloj's internal FCS list format.
}
\details{
The orloj FCS list format will accumulate more fields as analyses are
applied to it. For example, pre-processing will add a mask to find the
non-bead indices. You can use \code{\link{fcsExprs}} the get the expression
data after applying all of the masks that are in the list.
}
\seealso{
\code{\link{isSample}}, \code{\link{fcsExprs}}
}
