% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citeseq.R
\name{isFileCiteseq}
\alias{isFileCiteseq}
\title{Verify File is CITEseq.}
\usage{
isFileCiteseq(filename, min_adt = 5, min_gex = 500)
}
\arguments{
\item{filename}{The name of the file to test.}

\item{min_adt}{Minimum number of ADT columns.}

\item{min_gex}{Minimum number of GEX columns.}
}
\value{
TRUE if file is CITEseq, FALSE otherwise.
}
\description{
Verify that a given file is a CITEseq file by examining the header and
looking for "cell_barcode", "adt_", and "gex_" values.
}
