% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{jsDivergence}
\alias{jsDivergence}
\title{Jensen-Shannon Divergence}
\usage{
jsDivergence(p, q)
}
\arguments{
\item{p, q}{Density vectors of a one-dimensional probability distribution.}
}
\value{
JS divergence between p and q.
}
\description{
Compute the Jensen-Shannon (JS) divergence between two one-dimensional
probability distributions. Original code adopted from:
http://stackoverflow.com/questions/11226627/jensen-shannon-divergence-in-r
}
