% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{jsDivergenceMtx}
\alias{jsDivergenceMtx}
\title{Jensen-Shannon Divergence Between Matrices}
\usage{
jsDivergenceMtx(mtx1, mtx2, n = 25)
}
\arguments{
\item{mtx1, mtx2}{Numeric matrices which provide the data for the JS
divergence calculation.}

\item{n}{Number of grid points in each direction. Can be a scalar or a
length-2 integer vector.}
}
\value{
JS divergence between mtx1 and mtx2.
}
\description{
Compute the Jensen-Shannon (JS) divergence between two matrices. The matrices
are each converted into a one-dimensional probability distribution which are
then feed into the JS equation.
}
