% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{kde2dVec}
\alias{kde2dVec}
\title{Vectorized Two-dimensional Kernel Density Estimation}
\usage{
kde2dVec(mtx, n = 25, lims = c(range(mtx[, 1], mtx[, 2])))
}
\arguments{
\item{mtx}{A numeric matrix which provides the data for the kernel density
estimation.}

\item{n}{Number of grid points in each direction. Can be a scalar or a
length-2 integer vector.}

\item{lims}{The limits of the rectangle covered by the grid as
c(xl, xu, yl, yu).}
}
\value{
A vector for the one-dimensional probability distribution.
}
\description{
A wrapper around \code{\link[MASS]{kde2d}} that converts the density
estimation into a one-dimensional probability distribution.
}
