% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{loadSample}
\alias{loadSample}
\title{Load Astrolabe sample.}
\usage{
loadSample(experiment, sample_id = NULL, sample_name = NULL)
}
\arguments{
\item{experiment}{An Astrolabe experiment.}

\item{sample_id}{Astrolabe sample ID.}

\item{sample_name}{Astrolabe sample name.}
}
\value{
Sample list with analysis data.
}
\description{
Load FCS data and any additional analysis that has been exported by the
Astrolabe platform. Sample is identified through ID or name.
}
