% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_cytometry.R
\name{massFindNonBeadEvents}
\alias{massFindNonBeadEvents}
\title{Find the indices of non-bead events.}
\usage{
massFindNonBeadEvents(sample, bead_percentile, min_n = 200)
}
\arguments{
\item{sample}{An Astrolabe sample.}

\item{bead_percentile}{Clusters where at least one of the bead channels is
below this threshold are non-bead clusters.}

\item{min_n}{Minimum number of events in order to try and detect beads.}
}
\value{
Sample, with additional non_bead_indices and non_bead_message.
fields.
}
\description{
Non-bead event detection first clusters the data using the FlowSOM algorithm.
Then, for each cluster, the mean of all bead channels is calculated. Any
cluster that has at least one mean channel value below the bead_percentile
for that channel is a non-bead cluster. The indices of all cells in non-bead
clusters are non-bead events.
}
