% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_cytometry.R
\name{massPreprocess}
\alias{massPreprocess}
\title{Preprocess mass cytometry sample.}
\usage{
massPreprocess(sample, cofactor, bead_percentile)
}
\arguments{
\item{sample}{An Astrolabe sample.}

\item{cofactor}{Cofactor for asinh transformation.}

\item{bead_percentile}{Clusters where at least one of the bead channels is
below this threshold are non-bead clusters.}
}
\value{
Sample after the above steps are done.
}
\description{
Quality control, cleaning, and transformation for mass cytometry sample.
}
