% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_cytometry.R
\name{massTransformMassChannels}
\alias{massTransformMassChannels}
\title{Transform expression data for mass cytometry sample.}
\usage{
massTransformMassChannels(sample, cofactor = 5)
}
\arguments{
\item{sample}{An Astrolabe sample.}

\item{cofactor}{Cofactor for asinh transformation.}
}
\value{
Sample, with expression data transformed.
}
\description{
Choose which channels to transform and apply the asinh transformation. Mass
cytometry channels are identified using the naming standard from the Fluidigm
software.
}
