% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nameVector}
\alias{nameVector}
\title{Name elements of vector with their value}
\usage{
nameVector(v)
}
\arguments{
\item{v}{A vector to be named.}
}
\value{
Same vector, with names set to values.
}
\description{
Name all of the elements of a vector with their value. This is useful in
conjunction with lapply, as it allows you to operate on the values of the
vector and naming the list with these values.
}
