% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotBarPlot}
\alias{plotBarPlot}
\title{Bar plot object.}
\usage{
plotBarPlot(data, x, y, fill = NULL, title = NULL,
  scale_y_labels = NULL, theme = NULL)
}
\arguments{
\item{data}{Dataset to use for the plot. If sample, exprs will be used.}

\item{x, y}{Column names for the X-axis and Y-axis, respectively.}

\item{fill}{Column name for bar fill.}

\item{title}{Plot title.}

\item{scale_y_labels}{Scaling function for Y-axis tick labels.}

\item{theme}{Modifications to the default ggplot theme.}
}
\value{
An orloj plot list with the plot object and any other parameters that
are required to export it.
}
\description{
Generates a bar plot (using geom_col) ggplot object.
}
