% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotBoxPlot}
\alias{plotBoxPlot}
\title{Box plot object.}
\usage{
plotBoxPlot(data, x, y, title = NULL, scale_y_labels = NULL,
  theme = NULL)
}
\arguments{
\item{data}{Dataset to use for the plot. If sample, exprs will be used.}

\item{x, y}{Column names for X-axis and Y-axis, respectively.}

\item{title}{Plot title.}

\item{scale_y_labels}{Scaling function for Y-axis tick labels.}

\item{theme}{Modifications to the default ggplot theme.}
}
\value{
An orloj plot list with the plot object and any other parameters that
are required to export it.
}
\description{
Box plot object.
}
