% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Heat map object.}
\usage{
plotHeatmap(hm, x, y, value, type = NULL, title = NULL,
  x_axis_order = NULL, y_axis_order = NULL, theme = NULL)
}
\arguments{
\item{hm}{Data frame with heatmap data.}

\item{x, y}{Column names for X-axis and Y-axis, respectively.}

\item{value}{Column name for tile values.}

\item{type}{Heatmap type. Accepted values are NULL, "cluster_labels", and
"abundance". This will set some of the heatmap's formatting.}

\item{title}{Plot title.}

\item{x_axis_order, y_axis_order}{Order of X- and Y-axis tick labels. If none
specified, \code{\link[gtools]{mixedsort}} will be used.}

\item{theme}{Modifications to the default ggplot theme.}
}
\value{
An orloj plot list with the plot object and any other parameters that
are required to export it.
}
\description{
Heat map object.
}
