% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotScatterPlot}
\alias{plotScatterPlot}
\title{Scatter plot object.}
\usage{
plotScatterPlot(data, x, y, color = NULL, alpha = 0.5, size = 1,
  xlim = NULL, ylim = NULL, title = NULL, theme = NULL,
  force_data = FALSE)
}
\arguments{
\item{data}{Dataset to use for the plot. If sample, exprs will be used.}

\item{x, y}{Column names for X-axis and Y-axis, respectively.}

\item{color}{Column name for the color aesthetic.}

\item{alpha}{geom_point alpha aesthetic.}

\item{size}{geom_point size aesthetic.}

\item{xlim}{x-axis limits.}

\item{ylim}{y-axis limits.}

\item{title}{Plot title.}

\item{theme}{Modifications to the default ggplot theme.}

\item{force_data}{If false, plots with more than 100 points won't have plot
data included.}
}
\value{
An orloj plot list with the plot object and any other parameters that
are required to export it.
}
\description{
Generates a biaxial scatter plot ggplot object from an Astrolabe sample or
a data frame.
}
