% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess an Astrolabe sample.}
\usage{
preprocess(sample, cofactor = 5, bead_percentile = 0.98)
}
\arguments{
\item{sample}{An Astrolabe sample.}

\item{cofactor}{Cofactor for asinh transformation.}

\item{bead_percentile}{Clusters where at least one of the bead channels is
below this threshold are non-bead clusters.}
}
\value{
Sample after the above steps are done.
}
\description{
Preprocess an Astrolabe sample.
}
