% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCellAssignments.R
\name{reportCellAssignments}
\alias{reportCellAssignments}
\title{Cell assignments report.}
\usage{
reportCellAssignments(sample, channels = NULL)
}
\arguments{
\item{sample}{An Astrolabe sample.}

\item{channels}{List of channels to include in report.}
}
\description{
Generate plots and CSV files that report on cell assignments to cell subsets.
For each level, the report includes the following over all (subset, channel)
combinations:
- Heatmap of channel median in that subset
- Heatmap of channel coefficient of variation in that subset
- Joy plots of channel distribution in each subset
}
