% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCellCounts.R
\name{reportCellCounts}
\alias{reportCellCounts}
\title{Cell counts report.}
\usage{
reportCellCounts(sample)
}
\arguments{
\item{sample}{An Astrolabe samlpe.}
}
\description{
Generate plots and CSV files that report on cell counts over all cell
subsets. Report includes bar plots of cell counts by level.
}
