% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCombineAggregateStatistics.R
\name{reportCombineAggregateStatistics}
\alias{reportCombineAggregateStatistics}
\title{Combined aggregate statistics reports.}
\usage{
reportCombineAggregateStatistics(experiment)
}
\arguments{
\item{experiment}{An Astrolabe experiment.}
}
\description{
Generate plots and CSV files that report on aggregate statistics over all
experiment samples and all cell subsets.
}
