% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDifferentialAbundanceAnalysis.R
\name{reportDifferentialAbundanceAnalysis}
\alias{reportDifferentialAbundanceAnalysis}
\title{Differential abundance analysis report.}
\usage{
reportDifferentialAbundanceAnalysis(experiment)
}
\arguments{
\item{experiment}{An Astrolabe experiment.}
}
\description{
Generate plots and CSV files that report on differential abundance analysis
over multiple samples, sample features, and cell subsets. For each sample
feature, the report includes spreadsheet and volcano plot of analysis
results.
}
