% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDifferentialExpressionAnalysis.R
\name{reportDifferentialExpressionAnalysis}
\alias{reportDifferentialExpressionAnalysis}
\title{Differential expression analysis report.}
\usage{
reportDifferentialExpressionAnalysis(experiment)
}
\arguments{
\item{experiment}{An Astrolabe experiment.}
}
\description{
Generate plots and CSV files that report on differential expression analysis
over multiple samples, sample features, and cell subsets.
}
