% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportMarkerThresholds.R
\name{reportMarkerThresholds}
\alias{reportMarkerThresholds}
\title{Marker threshold report.}
\usage{
reportMarkerThresholds(sample, max_n = 1e+05)
}
\arguments{
\item{sample}{An Astrolabe sample.}

\item{max_n}{Subsample the data to a maximum of max_n cells.}
}
\value{
An orloj report list with all of the required objects.
}
\description{
Generate plot for the positive and negative cells of each marker.
}
