% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportMds.R
\name{reportMds}
\alias{reportMds}
\title{MDS Map report.}
\usage{
reportMds(experiment)
}
\arguments{
\item{experiment}{An Astrolabe experiment.}
}
\description{
Generate plots and CSV files that report on the MDS map.
}
