% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportNonCells.R
\name{reportNonCells}
\alias{reportNonCells}
\title{Non-cells report.}
\usage{
reportNonCells(sample)
}
\arguments{
\item{sample}{An Astrolabe sample.}
}
\value{
An orloj report list with all of the required objects.
}
\description{
Generates plots that report on events with extreme FSC-A values, high FSC-W
values, or high LD values.'
}
