% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sampleCellSubsetChannelStatistics}
\alias{sampleCellSubsetChannelStatistics}
\title{Cell subset channel statistics.}
\usage{
sampleCellSubsetChannelStatistics(sample, level = "Assignment")
}
\arguments{
\item{sample}{An Astrolabe sample.}

\item{level}{Cell subset level. Currently supported levels are "Assignment"
and "Profiling".}
}
\value{
Cell subset channel intensity statistics for that level.
}
\description{
Given an Astrolabe sample, return the cell subset channel intensity
statistics for a labeling level.
}
