% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sampleCellSubsetCounts}
\alias{sampleCellSubsetCounts}
\title{Cell subset counts.}
\usage{
sampleCellSubsetCounts(sample, level = "Assignment")
}
\arguments{
\item{sample}{An Astrolabe sample.}

\item{level}{Cell subset level, such as "Assignment" or "Profiling".}
}
\value{
Cell subset counts for that level.
}
\description{
Given an Astrolabe sample, return the cell subset counts for a labeling
level.
}
