% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_cytometry.R
\name{auroraPreprocess}
\alias{auroraPreprocess}
\title{Preprocess Aurora sample.}
\usage{
auroraPreprocess(sample)
}
\arguments{
\item{sample}{An Astrolabe sample.}
}
\value{
Sample after the above steps are done.
}
\description{
Quality control and cleaning for Aurora sample.
}
\details{
Transformation of Aurora data should be done before converting the flow frame
into an Astrolabe sample.
}
\seealso{
\code{\link{auroraTransformChannels}}
}
