% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{differentialAbundanceAnalysis}
\alias{differentialAbundanceAnalysis}
\title{Differential abundance analysis.}
\usage{
differentialAbundanceAnalysis(experiment, level = "Assignment")
}
\arguments{
\item{experiment}{An Astrolabe experiment.}

\item{level}{Cell subset level. Currently supported levels are "Assignment"
and "Profiling".}
}
\value{
Differential abundance analysis list.
}
\description{
Load the experiment differential abundance analysis, for a given cell subset
level.
}
