% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experimentCellSubsetChannelStatistics}
\alias{experimentCellSubsetChannelStatistics}
\title{Experiment cell subset channel statistics.}
\usage{
experimentCellSubsetChannelStatistics(experiment, level = "Assignment")
}
\arguments{
\item{experiment}{An Astrolabe experiment.}

\item{level}{Cell subset level. Currently supported levels are "Assignment"
and "Profiling".}
}
\value{
Cell subset channel statistics for that level.
}
\description{
Cell subset channel statistics for all of the samples in an experiment.
}
