% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experimentCellSubsetCounts}
\alias{experimentCellSubsetCounts}
\title{Experiment cell subset counts.}
\usage{
experimentCellSubsetCounts(experiment, level = .chooseLevel(experiment))
}
\arguments{
\item{experiment}{An Astrolabe experiment.}

\item{level}{Cell subset level, such as "Assignment" or "Profiling".}
}
\value{
Cell subset counts for that level.
}
\description{
Cell subset counts and frequencies for all of the samples in an experiment.
}
