% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experimentCellSubsetMap}
\alias{experimentCellSubsetMap}
\title{Map experiment cell subsets to numerical values.}
\usage{
experimentCellSubsetMap(experiment, level = .chooseLevel(experiment))
}
\arguments{
\item{experiment}{An Astrolabe experiment.}

\item{level}{Cell subset level. Currently supported levels are "Assignment"
and "Profiling". The default is "Profiling" for Profiling Only experiments
and "Assignment" otherwise.}
}
\value{
Map from cell subsets to numerical values.
}
\description{
Create a map from experiment cell subsets in a given level to unique
numerical values. This map is used when generating Assignment and Profiling
columns in exported FCS files.
}
