% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{fcsExprs}
\alias{fcsExprs}
\title{Get expression data from sample.}
\usage{
fcsExprs(sample, keep_debris = FALSE)
}
\arguments{
\item{sample}{An Astrolabe sample.}

\item{keep_debris}{Whether events in the Debris and Root/Unassigned labels
should be kept.}
}
\value{
Expression (and analyses) data frame.
}
\description{
Retrieves the exprs data frame from the sample, incorporating all relevant
pre-processing and Astrolabe analyses.
}
