% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{loadExperiment}
\alias{loadExperiment}
\title{Load Astrolabe experiment.}
\usage{
loadExperiment(experiment_path, astrolabe_path = NULL)
}
\arguments{
\item{experiment_path}{Path to the Astrolabe experiment.}

\item{astrolabe_path}{Internal parameter for Astrolabe Diagnostics staff.}
}
\value{
Astrolabe experiment.
}
\description{
Load Astrolabe experiment.
}
