% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{loadExperiment}
\alias{loadExperiment}
\title{Load Astrolabe experiment.}
\usage{
loadExperiment(experiment_path)
}
\arguments{
\item{experiment_path}{Path to the Astrolabe experiment.}
}
\value{
Astrolabe experiment.
}
\description{
Load an Astrolabe experiment which can then be used for other orloj
functions. As part of loading the experiment, correct the file paths that
were generated by the Astrolabe platform to local file paths.
}
