% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_cytometry.R
\name{massPreprocess}
\alias{massPreprocess}
\title{Preprocess mass cytometry sample.}
\usage{
massPreprocess(sample, cofactor = 5, bead_percentile = 0.99)
}
\arguments{
\item{sample}{An Astrolabe sample.}

\item{cofactor}{Cofactor for asinh transformation.}
}
\value{
Sample after the above steps are done.
}
\description{
Quality control, cleaning, and transformation for mass cytometry sample.
}
