% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess an Astrolabe sample.}
\usage{
preprocess(sample, cofactor = 5)
}
\arguments{
\item{sample}{An Astrolabe sample.}

\item{cofactor}{Cofactor for asinh transformation.}
}
\value{
Sample after the above steps are done.
}
\description{
Preprocess an Astrolabe sample.
}
