% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess an Astrolabe sample.}
\usage{
preprocess(experiment, sample)
}
\arguments{
\item{experiment}{An Astrolabe experiment.}

\item{sample}{An Astrolabe sample.}
}
\value{
Sample after the above steps are done.
}
\description{
Preprocess an Astrolabe sample.
}
