% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportBeads.R
\name{reportBeads}
\alias{reportBeads}
\title{Bead report.}
\usage{
reportBeads(sample)
}
\arguments{
\item{sample}{An Astrolabe sample.}
}
\value{
An orloj report list with all of the required objects.
}
\description{
Generates plots and CSV files that report on the beads in a mass cytometry
file.
}
