% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCellAssignments.R
\name{reportCellAssignments}
\alias{reportCellAssignments}
\title{Cell assignments report.}
\usage{
reportCellAssignments(sample, report_levels = FALSE)
}
\arguments{
\item{sample}{An Astrolabe sample.}

\item{report_levels}{If true, intermediate level heatmaps will be exported
in addition to terminal level heatmaps.}
}
\description{
Generate plots and CSV files that report on cell assignments to cell subsets.
Report includes a heatmap of all cell subsets in each level, and biaxial
plots of each subset separately.
}
