% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDifferentialExpressionAnalysis.R
\name{reportDifferentialExpressionAnalysis}
\alias{reportDifferentialExpressionAnalysis}
\title{Differential expression analysis report.}
\usage{
reportDifferentialExpressionAnalysis(
  experiment,
  kit_id,
  ridge_plots = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{experiment}{An Astrolabe experiment.}

\item{kit_id}{DE analysis kit ID.}

\item{ridge_plots}{Whether ridge plots should be exported.}
}
\description{
Generate plots and CSV files that report on differential expression analysis
for a single kit.
}
