% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportNonCells.R
\name{reportNonCells}
\alias{reportNonCells}
\title{Non-cells report.}
\usage{
reportNonCells(sample, max_n = Inf)
}
\arguments{
\item{sample}{An Astrolabe sample.}

\item{max_n}{Subsample the data to a maximum of max_n cells.}
}
\value{
An orloj report list with all of the required objects.
}
\description{
Generate plots for debris, doublets, and live/dead events.
}
